Changed type of primary key from User and Category to long and adapted methods

+++ src/main/java/com/lhama/lhamapersonalfinances/controllers/AuthenticationController.java
@@ -50,7 +50,7 @@ public class AuthenticationController {
 
         if (newUser == null) return ResponseEntity.internalServerError().build();
 
-        Integer idNewUser = newUser.getIdUser();
+        Long idNewUser = newUser.getIdUser();
 
         URI uri = uriBuilder.path("auth/register/{id}").buildAndExpand(idNewUser).toUri();
